+++ src/index.ts
@@ -1,11 +1,17 @@
 export type {Task, TaskContext} from './task/task.js';
-export {TaskError} from './task/task.js';
-
-export type {Gen} from './gen/gen.js';
+export type {Gen, GenContext} from './gen/gen.js';
 
 // by definition, these are generic, so just export everything
 export * from './utils/types.js';
 
-// these seem useful and generic enough to
+// these seem useful and generic enough to export to users
 export type {AsyncStatus} from './utils/async.js';
+export type {SpawnedProcess, SpawnResult} from './utils/process.js';
+export type {Lazy} from './utils/function.js';
+export type {ErrorClass} from './utils/error.js';
+
+// types above, code below
 export {wait, wrap} from './utils/async.js';
+export {last, toArray, EMPTY_ARRAY} from './utils/array.js';
+export {loadPackageJson} from './project/packageJson.js';
+export {TaskError} from './task/task.js';